# build: use enry to detect low priority files