--- a/build/e2e_test.go
+++ b/build/e2e_test.go
@@ -502,7 +502,7 @@ func TestFileRank(t *testing.T) {
 		name: "test",
 		docs: []*zoekt.Document{
 			{
-				Name:    "test",
+				Name:    "foo_test.go",
 				Content: []byte("bla"),
 			},
 			{