--- a/internal/e2e/testdata/time_compare.txt
+++ b/internal/e2e/testdata/time_compare.txt
@@ -29,10 +29,10 @@ github.com/golang/go/src/go/constant/value.go
 1381:		re := Compare(x.re, token.EQL, y.re)
 hidden 1 more line matches
 
-github.com/golang/go/src/syscall/zsyscall_windows.go
-878:func GetSystemTimeAsFileTime(time *Filetime) {
-1088:func SetFileTime(handle Handle, ctime *Filetime, atime *Filetime, wtime *Filetime) (err error) {
-132:	procGetSystemTimeAsFileTime            = modkernel32.NewProc("GetSystemTimeAsFileTime")
-hidden 19 more line matches
+github.com/golang/go/src/cmd/go/internal/gover/gover.go
+36:func Compare(x, y string) int {
+20:// but at the time this code was written, there was an existing test that used
+49:	if c := cmp.Compare(vx.kind, vy.kind); c != 0 { // "" < alpha < beta < rc
+hidden 4 more line matches
 
 hidden 139 more file matches