This is a much more robust detection mechanism. Additionally we have these signals we can also add in:

  func IsConfiguration(path string) bool
  func IsDocumentation(path string) bool
  func IsDotFile(path string) bool
  func IsImage(path string) bool

My main concern with this change is generated file detection on content using up RAM or CPU. Will monitor this impact on pprof in production.

Test Plan: go test.