@keegancsmith 2024-09-19
I did not! I confess to being lazy here since I don't have that repo cloned on my laptop (and my desktop is waiting for a new harddrive to be delivered). To be honest I thought this was low risk enough to monitor the impact on dotcom, but I realise that may actually not that be as clear given you are also shipping memory use improvements. I'll do a quick eval on the sg repo.